feat: naive impl on user signup endpoint
A naive version of the `signup` service is implemented,
providing the capability of registering users in the platform.

Its important to note that theres a lot of work to be done in
this feature, including:

- e2e testing
- Validations on incoming payload
- Code cleaning
- Error handling
- Configure SQLx for offline mode to make workflows run

And perhaps theres more points I will encounter for the moment
I work on those missing items.

I think is important to merge this version as this establishes a
project scaffold for all of us to contribute/expand this project.

Example

In order to register a new user you must send an HTTP request
which looks like the following:

```
POST /signup HTTP/1.1
Content-Type: application/json; charset=utf-8
Host: 0.0.0.0:7878
Connection: close
Content-Length: 101

{"email":"eborai@hilow.com","username":"eborai","name":"Esteban","surname":"Borai","password":"1234"}
```